+++ Source/AllOverIt.Aws.Cdk.AppSync/AppGraphqlProps.cs
@@ -1,4 +1,5 @@
-﻿using Amazon.CDK.AWS.AppSync;
+﻿using AllOverIt.Aws.Cdk.AppSync.Attributes.DataSources;
+using Amazon.CDK.AWS.AppSync;
 using Cdklabs.AwsCdkAppsyncUtils;
 using System.Collections.Generic;
 
@@ -7,6 +8,7 @@ namespace AllOverIt.Aws.Cdk.AppSync
     /// <summary>Contains options for an AppSync GraphQL API with the Schema pre-configured as a <see cref="CodeFirstSchema"/>.</summary>
     public sealed class AppGraphqlProps : GraphqlApiProps
     {
+        /// <summary>Provides endpoint lookup values for <see cref="HttpDataSourceAttribute"/> using <see cref="EndpointSource.Lookup"/>.</summary>
         public IReadOnlyDictionary<string, string> EndpointLookup { get; init; } = new Dictionary<string, string>();
 
         /// <summary>Constructor.</summary>